TradingBot/tests/test_models.py
"""Unit tests for ML models."""

import pytest
import pandas as pd
import numpy as np
from unittest.mock import Mock, patch
from sklearn.metrics import accuracy_score, r2_score

from models.signal_model import SignalModel
from config.config_loader import ConfigLoader


class TestSignalModel:
    """Test ML signal generation models."""
    
    @pytest.fixture
    def config(self):
        """Create test configuration."""
        config = Mock(spec=ConfigLoader)
        config.get = Mock(side_effect=lambda key, default=None: {
            'ml_models.classifier.target_accuracy': 0.70,
            'ml_models.classifier.min_confidence': 0.80,
            'ml_models.regressor.target_r2': 0.50,
            'ml_models.regressor.target_rmse': 2.0,
        }.get(key, default))
        return config
    
    @pytest.fixture
    def sample_data(self):
        """Create sample training data."""
        np.random.seed(42)
        n_samples = 1000
        
        # Generate features
        df = pd.DataFrame({
            'close': np.random.uniform(1000, 1100, n_samples),
            'rsi_14': np.random.uniform(30, 70, n_samples),
            'macd': np.random.uniform(-10, 10, n_samples),
            'ema_9': np.random.uniform(1000, 1100, n_samples),
            'ema_21': np.random.uniform(1000, 1100, n_samples),
            'atr_14': np.random.uniform(10, 30, n_samples),
            'volume': np.random.uniform(1000000, 5000000, n_samples),
        })
        
        # Generate labels (for classifier: win = 1, loss = 0)
        # Correlated with RSI and MACD
        df['win'] = ((df['rsi_14'] > 50) & (df['macd'] > 0)).astype(int)
        
        # Generate returns (for regressor)
        # Correlated with RSI
        df['return_pct'] = (df['rsi_14'] - 50) * 0.1 + np.random.normal(0, 1, n_samples)
        
        return df
    
    @pytest.fixture
    def model(self, config):
        """Create SignalModel instance."""
        return SignalModel(config)
    
    def test_model_initialization(self, model):
        """Test model initializes correctly."""
        assert model.classifier is None
        assert model.regressor is None
        assert model.feature_columns is None
    
    def test_train_classifier(self, model, sample_data):
        """Test classifier training."""
        features = sample_data[['rsi_14', 'macd', 'ema_9', 'ema_21', 'atr_14', 'volume']]
        labels = sample_data['win']
        
        metrics = model.train_classifier(features, labels)
        
        # Check model was trained
        assert model.classifier is not None
        
        # Check metrics
        assert 'accuracy' in metrics
        assert 'precision' in metrics
        assert 'recall' in metrics
        assert 'f1_score' in metrics
        
        # Should achieve reasonable accuracy on this simple dataset
        assert metrics['accuracy'] > 0.50
    
    def test_train_regressor(self, model, sample_data):
        """Test regressor training."""
        features = sample_data[['rsi_14', 'macd', 'ema_9', 'ema_21', 'atr_14', 'volume']]
        returns = sample_data['return_pct']
        
        metrics = model.train_regressor(features, returns)
        
        # Check model was trained
        assert model.regressor is not None
        
        # Check metrics
        assert 'r2_score' in metrics
        assert 'rmse' in metrics
        assert 'mae' in metrics
        
        # Should achieve reasonable fit
        assert metrics['r2_score'] > -1.0  # At least better than random
    
    def test_predict_without_training(self, model, sample_data):
        """Test prediction fails if model not trained."""
        features = sample_data[['rsi_14', 'macd', 'ema_9', 'ema_21', 'atr_14', 'volume']].iloc[:1]
        
        with pytest.raises(ValueError):
            model.predict(features)
    
    def test_predict_with_trained_models(self, model, sample_data):
        """Test prediction with trained models."""
        # Train models
        features = sample_data[['rsi_14', 'macd', 'ema_9', 'ema_21', 'atr_14', 'volume']]
        model.train_classifier(features, sample_data['win'])
        model.train_regressor(features, sample_data['return_pct'])
        
        # Predict on new data
        test_features = features.iloc[:5]
        predictions = model.predict(test_features)
        
        assert 'p_win' in predictions
        assert 'expected_return' in predictions
        assert 'confidence' in predictions
        assert 'should_trade' in predictions
        
        # Check value ranges
        assert 0 <= predictions['p_win'] <= 1
        assert predictions['confidence'] >= 0
        assert isinstance(predictions['should_trade'], bool)
    
    def test_predict_confidence_threshold(self, model, sample_data, config):
        """Test that predictions respect confidence threshold."""
        # Train models
        features = sample_data[['rsi_14', 'macd', 'ema_9', 'ema_21', 'atr_14', 'volume']]
        model.train_classifier(features, sample_data['win'])
        model.train_regressor(features, sample_data['return_pct'])
        
        # Set high confidence threshold
        config.get = Mock(return_value=0.95)  # Very high threshold
        
        # Most predictions should not meet threshold
        test_features = features.iloc[:100]
        predictions_list = [model.predict(features.iloc[[i]]) for i in range(100)]
        
        should_trade_count = sum(p['should_trade'] for p in predictions_list)
        
        # With 95% threshold, very few should pass
        assert should_trade_count < 50  # Less than half
    
    def test_feature_importance(self, model, sample_data):
        """Test feature importance extraction."""
        features = sample_data[['rsi_14', 'macd', 'ema_9', 'ema_21', 'atr_14', 'volume']]
        model.train_classifier(features, sample_data['win'])
        
        importance = model.get_feature_importance()
        
        assert len(importance) == len(features.columns)
        
        # Check that RSI and MACD are important (since labels are based on them)
        assert 'rsi_14' in importance
        assert 'macd' in importance
        
        # Importance should sum to ~1.0
        total_importance = sum(importance.values())
        assert 0.95 <= total_importance <= 1.05
    
    def test_save_and_load_models(self, model, sample_data, tmp_path):
        """Test model persistence."""
        # Train models
        features = sample_data[['rsi_14', 'macd', 'ema_9', 'ema_21', 'atr_14', 'volume']]
        model.train_classifier(features, sample_data['win'])
        model.train_regressor(features, sample_data['return_pct'])
        
        # Save
        save_dir = str(tmp_path / "models")
        model.save_models(save_dir, "test_model")
        
        # Create new model and load
        new_model = SignalModel(model.config)
        new_model.load_models(save_dir, "test_model")
        
        # Check models loaded
        assert new_model.classifier is not None
        assert new_model.regressor is not None
        assert new_model.feature_columns == model.feature_columns
        
        # Check predictions match
        test_features = features.iloc[:5]
        original_pred = model.predict(test_features)
        loaded_pred = new_model.predict(test_features)
        
        assert abs(original_pred['p_win'] - loaded_pred['p_win']) < 0.01
        assert abs(original_pred['expected_return'] - loaded_pred['expected_return']) < 0.01
    
    def test_model_retraining(self, model, sample_data):
        """Test that model can be retrained with new data."""
        features = sample_data[['rsi_14', 'macd', 'ema_9', 'ema_21', 'atr_14', 'volume']]
        
        # Train first time
        metrics1 = model.train_classifier(features, sample_data['win'])
        
        # Train again (should overwrite)
        metrics2 = model.train_classifier(features, sample_data['win'])
        
        # Should work without errors
        assert model.classifier is not None
        assert 'accuracy' in metrics2
    
    def test_prediction_with_missing_features(self, model, sample_data):
        """Test prediction fails gracefully with missing features."""
        # Train with all features
        features = sample_data[['rsi_14', 'macd', 'ema_9', 'ema_21', 'atr_14', 'volume']]
        model.train_classifier(features, sample_data['win'])
        model.train_regressor(features, sample_data['return_pct'])
        
        # Try to predict with missing feature
        incomplete_features = sample_data[['rsi_14', 'macd']].iloc[:1]
        
        with pytest.raises((ValueError, KeyError)):
            model.predict(incomplete_features)
    
    def test_prediction_batch(self, model, sample_data):
        """Test batch prediction."""
        # Train models
        features = sample_data[['rsi_14', 'macd', 'ema_9', 'ema_21', 'atr_14', 'volume']]
        model.train_classifier(features, sample_data['win'])
        model.train_regressor(features, sample_data['return_pct'])
        
        # Predict on batch
        batch_features = features.iloc[:10]
        
        # Should work for batch
        # Note: Current implementation returns single dict, 
        # so this test checks if it handles batches at all
        try:
            predictions = model.predict(batch_features)
            # If it works, great
            assert predictions is not None
        except:
            # If not implemented, that's also acceptable
            pass


class TestModelTraining:
    """Test model training pipeline."""
    
    def test_label_generation_winning_trades(self):
        """Test label generation for winning trades."""
        # Create sample price data with clear winning pattern
        prices = pd.Series([100, 105, 110, 108, 115])  # Net uptrend
        
        # Labels should be 1 (win) for entries that lead to profit
        # Implementation depends on your label generation logic
        pass
    
    def test_label_generation_losing_trades(self):
        """Test label generation for losing trades."""
        # Create sample price data with clear losing pattern
        prices = pd.Series([100, 95, 90, 92, 88])  # Net downtrend
        
        # Labels should be 0 (loss)
        pass
    
    def test_feature_scaling(self):
        """Test that features are properly scaled if needed."""
        # Some models benefit from feature scaling
        # Test if your implementation does this
        pass


if __name__ == "__main__":
    pytest.main([__file__, "-v"])
