TradingBot/requirements.txt
# Core Python packages
python-dotenv==1.0.0

# Data processing
pandas==2.1.3
numpy==1.26.2
pyarrow==14.0.1

# ML frameworks
scikit-learn==1.3.2
xgboost==2.0.2
catboost==1.2.2

# Technical analysis
# TA-Lib==0.4.28  # Requires C library installation - see INSTALL_TALIB.md
# Alternative: Use ta library instead
ta==0.11.0

# Async & networking
aiohttp==3.9.1
asyncio==3.4.3
websockets==12.0

# Configuration & validation
pyyaml==6.0.1
pydantic==2.5.2
pydantic-settings==2.1.0

# Zerodha Kite Connect
kiteconnect==4.2.0
pyotp==2.9.0  # For TOTP authentication

# Data storage
redis==5.0.1
influxdb-client==1.38.0

# Monitoring & logging
prometheus-client==0.19.0
loguru==0.7.2

# Visualization (optional)
matplotlib==3.8.2
plotly==5.18.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
mock==5.1.0

# Utilities
python-dateutil==2.8.2
requests==2.31.0
